Change header to module
Note: There's an issue with CMake that causes the object file from the
static library to have the wrong name. This only happens with modules
and MSVC. The workaround is to manually create a copy of the object file
and give it the correct name (i.e., remove the ".ixx" portion).